+++ src/kits/media/MediaTrack.cpp
@@ -530,7 +530,7 @@ BMediaTrack::ReadChunk(char** _buffer, int32* _size, media_header* _header)
 		// not work so well because of the chunk cache. But providing a
 		// "duration" field in the media_header could be useful.
 		fCurrentTime = fCurrentTime = _header->start_time;
-		fCurrentFrame = fCurrentTime * _FrameRate() / 1000000LL;
+		fCurrentFrame = (int64)(fCurrentTime * _FrameRate() / 1000000LL);
 
 	}
